ISSR_Communication_Analysis_Tool_Samuel_Kalu/requirements.txt
moviepy==1.0.3
transformers==4.35.2 
torchaudio==2.7.1
librosa==0.10.1 
spacy==3.8.7
pandas==2.1.0 
matplotlib==3.8.0 
seaborn==0.13.0
openai-whisper==20240930
gradio==5.23.1
pyannote.audio==3.3.2
loguru==0.7.3
vaderSentiment==3.3.2
faster-whisper==1.1.1
gdown==5.2.0